Finished v1

+++ index.js
@@ -76,6 +76,10 @@ client.on('message', async (m) => {
         if (m.author.id in tagEloQueues) return m.channel.send("You are already queued")
         if (m.author.id in tagEloGames) return m.channel.send(`You stil have a game against <@!${tagEloGames[m.author.id]}>`)
 
+        if (!m.author.id in await db.get(`data`)) {
+            return m.channel.send("You must be verified first. Contact a Staff member for support")
+        }
+
         setTimeout(async function () {
             m.delete()
         }, 100);
@@ -174,11 +178,79 @@ client.on('message', async (m) => {
         })
     }
     else if (command == "stats") {
+
+        if (!m.author.id in await db.get(`data`)) {
+            return m.channel.send("You must be verified first. Contact a Staff member for support")
+        }
         return m.channel.send(`
 Rating: ${await db.get(`data.${m.author.id}.rating`)}
 Wins: ${await db.get(`data.${m.author.id}.performace.wins`)}
 Losses: ${await db.get(`data.${m.author.id}.performace.losses`)}
 Estimated Performance: ${await calculatePerformance(m.author.id)}`)
+    }
+    else if (command == "leaderboard") {
+        let data = Object.entries(await db.get(`data`))
+        data = data.sort((a, b) => {return b[1].rating - a[1].rating})
+
+        let position = null;
+        if (m.author.id in await db.get(`data`)) { 
+            for (let i = 0; i<data.length; i++) {
+                if (data[i][0] == m.author.id) {
+                    position = i+1
+                    break;
+                }
+            }
+        }
+
+        return m.channel.send(`**ELO Leaderboard**
+${(position = 1) ? "**" : ""}#1 <@!${data[0][0]}> - ${data[0][1].rating}${(position = 1) ? "**" : ""}
+${(position = 2) ? "**" : ""}#2 <@!${data[1][0]}> - ${data[1][1].rating}${(position = 2) ? "**" : ""}
+${(position = 3) ? "**" : ""}#3 <@!${data[2][0]}> - ${data[2][1].rating}${(position = 3) ? "**" : ""}
+${(position = 4) ? "**" : ""}#4 <@!${data[3][0]}> - ${data[3][1].rating}${(position = 4) ? "**" : ""}
+${(position = 5) ? "**" : ""}#5 <@!${data[4][0]}> - ${data[4][1].rating}${(position = 5) ? "**" : ""}
+${(position = 6) ? "**" : ""}#6 <@!${data[5][0]}> - ${data[5][1].rating}${(position = 6) ? "**" : ""}
+${(position = 7) ? "**" : ""}#7 <@!${data[6][0]}> - ${data[6][1].rating}${(position = 7) ? "**" : ""}
+${(position = 8) ? "**" : ""}#8 <@!${data[7][0]}> - ${data[7][1].rating}${(position = 8) ? "**" : ""}
+${(position > 8) ? `. . .\n**#${position} <@!${data[position - 1][0]}> - ${data[position - 1][1].rating}` : ""}**`)
+    }
+    else if (command == "verify") {
+        if (m.member.roles.cache.has('752177298161008763')) return
+        if (m.mentions.members.size != 0) return m.channel.send("Ping the *one* person you would like to verify")
+
+        await db.set(`data.${m.mentions.members.first().id}`, {"rating":1500, "performance":{"wins":0, "losses":0, "opponents":[]}})
+        await db.set(`settings.${m.author.id}.deviation`, 100)
+    }
+    else if (command == "duel") {
+        if (!m.channel.id in config.mainChannelID) return
+        if (!m.author.id in await db.get(`data`)) {
+            return m.channel.send("You must be verified first. Contact a Staff member for support")
+        }
+        if (m.author.id in tagEloGames) return m.channel.send(`You are in a game against <@!${tagEloGames[m.author.id]}>. If you can't make it you must either forfeit (do ${prefix}winner <@!${tagEloGames[m.author.id]}>) or abort (do ${prefix}abort)`)
+        if (m.author.id in tagEloQueues) return m.channel.send("You are in the queue. Leave it to duel")
+        pingedMember = m.mentions.members.first()
+        if (pingedMember.id in tagEloGames) return m.channel.send("That player is already in a game")
+        if (pingedMember.id in tagEloQueues) return m.channel.send("That player is in the queue. They must leave it to duel")
+        if (!pingedMember.id in await db.get(`data`)) {
+            return m.channel.send("You must be verified first. Contact a Staff member for support")
+        }
+
+        const msg = await m.channel.send(`<@!${pingedMember.id}> React with a :thumbsup: if you would like to duel <@!${m.author.id}> (Rated ${await db.get(`data.${m.author.id}.rating`)})`)
+
+        const filter = (reaction, user) => (user.id === pingedMember.id)
+        const collector = msg.createReactionCollector(filter, {time: 10000})
+
+        msg.react('👍')
+
+        collector.on('collect', async (reaction, user) => {
+            tagEloGames[m.author.id] = pingedMember.id
+            tagEloGames[pingedMember.id] = m.author.id
+            msg.edit(`Game Starting:
+<@!${m.author.id}> (${await db.get(`data.${m.author.id}.rating`)}) vs <@!${pingedMember.id}> (${await db.get(`data.${pingedMember.id}.rating`)})`)
+        })
+
+        collector.on('end', () => {
+            m.edit("Duel Request Denied")
+        })
     }
 })
